+++ src/constants.js
@@ -140,3 +140,7 @@ export const requestStatusFilters = [
 ];
 
 export const pickSlipType = 'pick slip';
+
+export const DOMAIN_NAME = 'requests';
+
+export const APP_ICON_NAME = 'requests';