+++ src/utils.js
@@ -238,3 +238,21 @@ export function buildUrl(location, values) {
 
   return isEmpty(params) ? url : `${url}?${queryString.stringify(params)}`;
 }
+
+export function formatNoteReferrerEntityData(data) {
+  if (data) {
+    const {
+      entityName: name,
+      entityType: type,
+      entityId: id,
+    } = data;
+
+    return {
+      name,
+      type,
+      id,
+    };
+  }
+
+  return false;
+}